Migrate client speed simulation to the end of each training epoch
